Add a script to separate the latest blog posts from the contentlayer json into a separate json.
