+++ apps/www/package.json
@@ -11,7 +11,7 @@
     "lint": "next lint",
     "clean": "rimraf node_modules",
     "typecheck": "npm run content:build && tsc --noEmit",
-    "content:build": "contentlayer2 build",
+    "content:build": "contentlayer2 build && node scripts/generateLatestBlogPosts.mjs",
     "prettier": "prettier --write \"./{pages,components,lib,stores,styles,tests}/**/*.{ts,tsx,md,js,jsx,json}\"",
     "postbuild": "node ./internals/generate-sitemap.mjs"
   },
